Renamed "source" directory to "src"
