+++ apps/tui-agent/src/cli.tsx
