+++ apps/tui-agent/test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import chalk from 'chalk';
 import test from 'ava';
 import {render} from 'ink-testing-library';
-import App from './source/app.js';
+import App from './src/app.js';
 
 test('greet unknown user', t => {
 	const {lastFrame} = render(<App name={undefined} />);
