+++ apps/tui-agent/tsconfig.json
@@ -3,5 +3,5 @@
 	"compilerOptions": {
 		"outDir": "dist"
 	},
-	"include": ["source"]
+	"include": ["src"]
 }
